requirements.txt
pandasai==1.5.11
streamlit==1.30.0
numpy==1.24.1
langchain==0.1.1
google-generativeai
